pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sparv-sbx-moses"
version = "0.1.0"
description = "Moses plugin for Sparv"
readme = "README.md"
license.text = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Språkbanken Text", email = "sb-info@svenska.gu.se" },
]
dependencies = [
    "sparv-pipeline~=5.0",
]
entry-points."sparv.plugin" = { moses = "moses" }

[project.urls]
Homepage = "https://github.com/spraakbanken/sparv-sbx-moses/"

[tool.hatch]
build.include = ["/moses"]
